+++ src/rtc/rtc-ctl.h
@@ -11,8 +11,8 @@ bool rtc_is_alarm(void);
 uint32_t rtc_get_time(void);
 uint32_t rtc_get_alarm(void);
 
-void atomic_state_set(uint8_t data);
-uint8_t atomic_state_get(void);
+void atomic_state_set(uint16_t data);
+uint16_t atomic_state_get(void);
 
 void backup_data_set(uint32_t data);
 uint32_t backup_data_get(void);
